refactor(home): replace detail overlay with navigation-based PlanDetailScreen
The plan detail view is now a proper route (plan-detail/{turnIndex}/{sessionId})
that navigates like settings detail pages: push slide-in, PredictiveBackCard
for gesture-driven back. HomeContent no longer manages a detail overlay,
TimelineMode, or predictive back state. Pull-to-expand thinking works via
ScrollState-based NestedScrollConnection in the new screen.
